Move all test files under the `/tests` directory
