+++ tests/test_external_services/test_hauki/test_hauki_resource_hash_updater.py
