+++ tests/test_external_services/test_verkkokauppa/test_merchant_types.py
