+++ tests/test_external_services/test_verkkokauppa/test_payment_types.py
