+++ tests/test_external_services/test_verkkokauppa/test_product_types.py
