+++ tests/test_external_services/test_verkkokauppa/test_webhooks/test_order_payment_webhooks.py
@@ -10,7 +10,7 @@ from merchants.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from reservations.choices import ReservationStateChoice
 from tests.factories import PaymentOrderFactory, ReservationFactory
 from tests.helpers import patch_method
-from tests.test_webhooks.helpers import get_mock_order_payment_api
+from tests.test_external_services.test_verkkokauppa.test_webhooks.helpers import get_mock_order_payment_api
 
 # Applied to all tests
 pytestmark = [
