+++ tests/test_external_services/test_verkkokauppa/test_webhooks/test_order_refund_webhooks.py
@@ -8,7 +8,7 @@ from merchants.verkkokauppa.payment.exceptions import GetRefundStatusError
 from merchants.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from tests.factories import PaymentOrderFactory
 from tests.helpers import patch_method
-from tests.test_webhooks.helpers import get_mock_order_refund_api
+from tests.test_external_services.test_verkkokauppa.test_webhooks.helpers import get_mock_order_refund_api
 
 # Applied to all tests
 pytestmark = [
