+++ tests/test_models/test_purpose_image.py
