+++ tests/test_tasks/test_prune_reservation_statistics.py
