+++ tests/test_utils/test_image_cache.py
